Monkey patch tests to make them skippable.
@puiterwijk and I cooked this up at Flock!

This makes it so that you can ctrl-C during any test and it will be
marked as SKIPPED instead of failing the whole test suite.  This is
especially nice if we have a service that we know happens to be broken
or if there's a service that is temporarily hung (or uninteresting or
whatever).

:bird: flocktofedora :bird: